Added code for writing the atom reference to a list

+++ MolecularRendererPackage/Sources/MolecularRenderer/Kernels/BVH/BVH+BuildSmall+Atoms.metal
@@ -190,7 +190,7 @@ kernel void buildSmallPart2_2
                                   1, memory_order_relaxed);
         
         // Write the reference to the list.
-        smallAtomReferences[offset] = uint(tid);
+        smallAtomReferences[offset] = tid;
       }
     }
   }
